Consolidate the OTAP proto definition in src/proto/mod.rs
