+++ rust/otel-arrow-rust/src/decode/record_message.rs
@@ -10,7 +10,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::opentelemetry::ArrowPayloadType;
+use crate::proto::opentelemetry::arrow::v1::ArrowPayloadType;
 use arrow::array::RecordBatch;
 
 /// Wrapper for [RecordBatch].
